# Delay the conversion to vim completion items