Partial fix for #291

A completion list can contain many more items than will be displayed if
many items are filled out due to not matching the already typed base
word. Any time spent performing translation from an LSP style completion
item to a vim style item is wasted for the words that aren't displayed.
Delay the conversion until after an item is known to not be filtered.

- Perform only the dict -> list normalization initially, and pass the
  list of LSP items to the next step of completion.
- Find the star of the completion using the LSP items.
- Pass the base string to `s:CompletionItems`, perform the filtering in
  same loop as the conversion. Skip any items which don't match the
  base.
- Split `s:CompletionItem` into `s:CompletionItemWord` which finds the
  text to insert, and optionally the snippet user data, and
  `s:FinishItem` which does the heavier normalization including
  documentation details.
- Add some smoke tests for autocomplete.