@natebosch 2020-06-07
It's not trivial to automate a large sample to get better benchmarks, bu ton a very small sample this looks promising.

Before:

```
FUNCTIONS SORTED ON TOTAL TIME
count  total (s)   self (s)  function
    2   0.830335   0.000057  <SNR>124_OnResult()
    2   0.687555   0.012475  <SNR>124_CompletionItems()
12417   0.656961   0.571318  <SNR>124_CompletionItem()
    2   0.142710   0.001329  <SNR>124_SuggestCompletions()
    1   0.141157   0.001483  <SNR>124_FindSuggestions()
    1   0.139674             <SNR>124_FilterAndSort()
 9236   0.085643             <SNR>124_CompletionItemKind()
    8   0.008859   0.000201  lsc#complete#textChanged()
    7   0.008658   0.000206  <SNR>124_typedCharacter()
    2   0.007773   0.000130  <SNR>124_startCompletion()
    7   0.000409   0.000256  <SNR>124_isTrigger()
    6   0.000270             <SNR>124_isCompletable()
    2   0.000110   0.000097  <SNR>124_SetCompleteOpt()
    3   0.000061   0.000031  lsc#complete#clean()
    9   0.000054             lsc#complete#insertCharPre()
    5   0.000043             <SNR>124_MarkNotCompleting()
    2   0.000011             <SNR>124_MarkCompleting()
    2   0.000008             <SNR>124_FindStart()
                             <SNR>124_OnSkip()
                             <SNR>124_GuessCompletionStart()
```

After:

```
FUNCTIONS SORTED ON TOTAL TIME
count  total (s)   self (s)  function
    2   0.338431   0.000070  <SNR>124_OnResult()
    2   0.338347   0.000497  <SNR>124_SuggestCompletions()
    2   0.337704   0.146500  <SNR>124_CompletionItems()
12417   0.189641             <SNR>124_CompletionItemWord()
    8   0.008709   0.000203  lsc#complete#textChanged()
    7   0.008506   0.000187  <SNR>124_typedCharacter()
    2   0.007642   0.000132  <SNR>124_startCompletion()
   39   0.001563   0.001410  <SNR>124_FinishItem()
    7   0.000410   0.000271  <SNR>124_isTrigger()
    6   0.000267             <SNR>124_isCompletable()
   24   0.000153             <SNR>124_CompletionItemKind()
    2   0.000060   0.000044  <SNR>124_SetCompleteOpt()
    3   0.000059   0.000030  lsc#complete#clean()
    9   0.000048             lsc#complete#insertCharPre()
    5   0.000043             <SNR>124_MarkNotCompleting()
    2   0.000023             <SNR>124_FindStart()
    2   0.000013             <SNR>124_MarkCompleting()
                             <SNR>124_OnSkip()
                             <SNR>124_GuessCompletionStart()
                             lsc#complete#complete()
```